requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
EditorConfig==0.12.4
Faker==30.8.2
fast-query-parsers==1.0.3
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jsbeautifier==1.15.1
litestar==2.12.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgspec==0.18.6
multidict==6.1.0
polyfactory==2.18.0
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pygame-ce==2.5.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
rich==13.9.4
rich-click==1.8.3
six==1.16.0
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.12.2
uvicorn==0.32.0
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.1
